change read only setup for ios
